Refix PcTester plus some changes to pc
Changed the test to byte addressing and took into account the previous
sugestions.
Changed the pc 'in_opcode' entrance of 10 bits to 2 separate of 7 bits
and 3 bits with the aporopriate changes in Core.scala
Fixed bug of 'program_counter' register being initiated as a 28 bits
register as opposed of 32 bits.